More porting to float
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/branches/libsidplayfp-experimental@2723 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/builders/residfp-builder/residfp/Integrator8580.h
@@ -136,7 +136,7 @@ int Integrator8580::solve(int vi) const
     vc += n_I_dac;
 
     // vx = g(vc)
-    const int tmp = (vc >> 15) + (1 << 15);
+    const float tmp = (vc >> 15) + (1 << 15);
     assert(tmp < (1 << 16));
     vx = opamp_rev->output(tmp);
